+++ src/binaryb16/support/check_exhaustive.c
@@ -183,9 +183,9 @@ doit (uint16_t n1, uint16_t n2)
   if (!is_equal (y, z))
   {
 #ifndef EXCHANGE_X_Y
-    printf ("FAIL x,y=%a,%a ref=%a z=%a\n", (float) x1, (float) x2, (float) y, (float) z);
+    printf ("FAIL x,y=%a,%a ref=%a z=%a\n", (double) x1, (double) x2, (double) y, (double) z);
 #else
-    printf ("FAIL y,x=%a,%a ref=%a z=%a\n", (float) x1, (float) x2, (float) y, (float) z);
+    printf ("FAIL y,x=%a,%a ref=%a z=%a\n", (double) x1, (double) x2, (double) y, (double) z);
 #endif
     fflush (stdout);
 #ifndef DO_NOT_ABORT
@@ -205,7 +205,7 @@ doit (uint16_t n1, uint16_t n2)
   // check spurious/missing underflow
   if (fetestexcept (FE_UNDERFLOW) && !mpfr_flags_test (MPFR_FLAGS_UNDERFLOW))
   {
-    printf ("Spurious underflow exception for x,y=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+    printf ("Spurious underflow exception for x,y=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
     fflush (stdout);
 #ifndef DO_NOT_ABORT
    	exit (1);
@@ -213,7 +213,7 @@ doit (uint16_t n1, uint16_t n2)
   }
   if (!fetestexcept (FE_UNDERFLOW) && mpfr_flags_test (MPFR_FLAGS_UNDERFLOW))
   {
-    printf ("Missing underflow exception for x,y=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+    printf ("Missing underflow exception for x,y=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
     fflush (stdout);
 #ifndef DO_NOT_ABORT
    	exit (1);
@@ -225,7 +225,7 @@ doit (uint16_t n1, uint16_t n2)
   // check spurious/missing overflow
   if (fetestexcept (FE_OVERFLOW) && !mpfr_flags_test (MPFR_FLAGS_OVERFLOW))
   {
-    printf ("Spurious overflow exception for x,y=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+    printf ("Spurious overflow exception for x,y=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
     fflush (stdout);
 #ifndef DO_NOT_ABORT
    	exit (1);
@@ -233,7 +233,7 @@ doit (uint16_t n1, uint16_t n2)
   }
   if (!fetestexcept (FE_OVERFLOW) && mpfr_flags_test (MPFR_FLAGS_OVERFLOW))
   {
-    printf ("Missing overflow exception for x,y=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+    printf ("Missing overflow exception for x,y=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
     fflush (stdout);
 #ifndef DO_NOT_ABORT
    	exit (1);
@@ -245,7 +245,7 @@ doit (uint16_t n1, uint16_t n2)
 #ifdef CORE_MATH_CHECK_INEXACT
   if ((inex_y == 0) && (inex_z != 0))
   {
-    printf ("Spurious inexact exception for x,y=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+    printf ("Spurious inexact exception for x,y=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
     fflush (stdout);
 #ifndef DO_NOT_ABORT
    	exit (1);
@@ -253,7 +253,7 @@ doit (uint16_t n1, uint16_t n2)
   }
   if ((inex_y != 0) && (inex_z == 0))
   {
-    printf ("Missing inexact exception for x,y=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+    printf ("Missing inexact exception for x,y=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
     fflush (stdout);
 #ifndef DO_NOT_ABORT
    	exit (1);
@@ -269,9 +269,9 @@ doit (uint16_t n1, uint16_t n2)
     if (is_nan (y) && errno != EDOM)
     {
 #ifndef EXCHANGE_X_Y
-      printf ("Missing errno=EDOM for x,y=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+      printf ("Missing errno=EDOM for x,y=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
 #else
-      printf ("Missing errno=EDOM for y,x=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+      printf ("Missing errno=EDOM for y,x=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
 #endif
       fflush (stdout);
 #ifndef DO_NOT_ABORT
@@ -280,7 +280,7 @@ doit (uint16_t n1, uint16_t n2)
     }
     if (!is_nan (y) && errno == EDOM)
     {
-      printf ("Spurious errno=EDOM for x,y=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+      printf ("Spurious errno=EDOM for x,y=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
       fflush (stdout);
 #ifndef DO_NOT_ABORT
       exit (1);
@@ -292,9 +292,9 @@ doit (uint16_t n1, uint16_t n2)
     if (expected_erange && errno != ERANGE)
     {
 #ifndef EXCHANGE_X_Y
-      printf ("Missing errno=ERANGE for x,y=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+      printf ("Missing errno=ERANGE for x,y=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
 #else
-      printf ("Missing errno=ERANGE for y,x=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+      printf ("Missing errno=ERANGE for y,x=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
 #endif
       fflush (stdout);
 #ifndef DO_NOT_ABORT
@@ -303,7 +303,7 @@ doit (uint16_t n1, uint16_t n2)
     }
     if (!expected_erange && errno == ERANGE)
     {
-      printf ("Spurious errno=ERANGE for x,y=%a,%a (z=%a)\n", (float) x1, (float) x2, (float) y);
+      printf ("Spurious errno=ERANGE for x,y=%a,%a (z=%a)\n", (double) x1, (double) x2, (double) y);
       fflush (stdout);
 #ifndef DO_NOT_ABORT
       exit (1);
@@ -318,7 +318,7 @@ doit (uint16_t n1, uint16_t n2)
   y = cr_function_under_test (x1, x2);
   if (!fetestexcept (FE_INEXACT)) {
     printf ("Exception inexact was reset for x,y=%a,%a (z=%a)\n",
-            (float) x1, (float) x2, (float) y);
+            (double) x1, (double) x2, (double) y);
     fflush (stdout);
 #ifndef DO_NOT_ABORT
     exit (1);
@@ -349,7 +349,7 @@ check_signaling_nan (void)
       if (!is_nan (y))
       {
       	fprintf (stderr, "Error, foo(sNaN=%x,y=%a) should be NaN, got %a=%x\n",
-                 u1, (float) x2, (float) y, asuint (y));
+                 u1, (double) x2, (double) y, asuint (y));
 #ifndef DO_NOT_ABORT
       	exit (1);
 #endif
@@ -358,7 +358,7 @@ check_signaling_nan (void)
       if (is_signaling (y))
       {
       	fprintf (stderr, "Error, foo(sNaN=%x,y=%a) should be qNaN, got sNaN=%x\n",
-               	 u1, (float) x2, asuint (y));
+               	 u1, (double) x2, asuint (y));
 #ifndef DO_NOT_ABORT
       	exit (1);
 #endif
@@ -368,7 +368,7 @@ check_signaling_nan (void)
       if (!is_nan (y))
       {
       	fprintf (stderr, "Error, foo(x=%a,sNaN=%x) should be NaN, got %a=%x\n",
-                 (float) x2, u1, (float) y, asuint (y));
+                 (double) x2, u1, (double) y, asuint (y));
 #ifndef DO_NOT_ABORT
       	exit (1);
 #endif
@@ -377,7 +377,7 @@ check_signaling_nan (void)
       if (is_signaling (y))
       {
       	fprintf (stderr, "Error, foo(x=%a,sNaN=%x) should be qNaN, got sNaN=%x\n",
-               	 (float) x2, u1, asuint (y));
+               	 (double) x2, u1, asuint (y));
 #ifndef DO_NOT_ABORT
       	exit (1);
 #endif
@@ -392,7 +392,7 @@ check_signaling_nan (void)
       if (!is_nan (y))
       {
       	fprintf (stderr, "Error, foo(sNaN=%x,y=%a) should be NaN, got %a=%x\n",
-                 u1, (float) x2, (float) y, asuint (y));
+                 u1, (double) x2, (double) y, asuint (y));
 #ifndef DO_NOT_ABORT
       	exit (1);
 #endif
@@ -401,7 +401,7 @@ check_signaling_nan (void)
       if (is_signaling (y))
       {
       	fprintf (stderr, "Error, foo(sNaN=%x,y=%a) should be qNaN, got sNaN=%x\n",
-               	 u1, (float) x2, asuint (y));
+               	 u1, (double) x2, asuint (y));
 #ifndef DO_NOT_ABORT
       	exit (1);
 #endif
@@ -411,7 +411,7 @@ check_signaling_nan (void)
       if (!is_nan (y))
       {
       	fprintf (stderr, "Error, foo(x=%a,sNaN=%x) should be NaN, got %a=%x\n",
-                 (float) x2, u1, (float) y, asuint (y));
+                 (double) x2, u1, (double) y, asuint (y));
 #ifndef DO_NOT_ABORT
       	exit (1);
 #endif
@@ -420,7 +420,7 @@ check_signaling_nan (void)
       if (is_signaling (y))
       {
       	fprintf (stderr, "Error, foo(x=%a,sNaN=%x) should be qNaN, got sNaN=%x\n",
-               	 (float) x2, u1, asuint (y));
+               	 (double) x2, u1, asuint (y));
 #ifndef DO_NOT_ABORT
       	exit (1);
 #endif
@@ -442,10 +442,10 @@ check_exceptions_aux (uint16_t n1, uint16_t n2)
   {
 #ifndef EXCHANGE_X_Y
     fprintf (stderr, "Error, for x,y=%a,%a, inexact exception set (z=%a)\n",
-             (float) x1, (float) x2, (float) y);
+             (double) x1, (double) x2, (double) y);
 #else
     fprintf (stderr, "Error, for y,x=%a,%a, inexact exception set (z=%a)\n",
-             (float) x1, (float) x2, (float) y);
+             (double) x1, (double) x2, (double) y);
 #endif
 #ifndef DO_NOT_ABORT
     exit (1);
@@ -457,7 +457,7 @@ check_exceptions_aux (uint16_t n1, uint16_t n2)
   if (inex)
   {
     fprintf (stderr, "Error, for x,y=%a,%a, overflow exception set (z=%a)\n",
-						 (float) x1, (float) x2, (float) y);
+						 (double) x1, (double) x2, (double) y);
 #ifndef DO_NOT_ABORT
     exit (1);
 #endif
@@ -468,7 +468,7 @@ check_exceptions_aux (uint16_t n1, uint16_t n2)
   if (inex)
   {
     fprintf (stderr, "Error, for x,y=%a,%a, underflow exception set (z=%a)\n",
-						 (float) x1, (float) x2, (float) y);
+						 (double) x1, (double) x2, (double) y);
 #ifndef DO_NOT_ABORT
     exit (1);
 #endif
